refactor(oauth): centralize redirect URI validation in worker-shared
Moves isValidRedirectUri() to @flaim/worker-shared so auth-worker and
web consent page use identical RFC 9700-compliant validation logic.

The web's isAllowedRedirectUri() was hostname-based and too permissive
(e.g. https://claude.ai/evil would pass), leaving the handleDeny() open
redirect unprotected. Both surfaces now use exact-match + structural
checks from the shared canonical implementation.

- workers/shared/src/oauth-redirect.ts: new canonical validation module
- workers/shared/src/index.ts: re-exports isValidRedirectUri
- workers/auth-worker/src/oauth-handlers.ts: imports from shared, removes local copy, re-exports for tests
- web/package.json: adds @flaim/worker-shared dependency
- web/tsconfig.json: path alias pointing to oauth-redirect source (avoids pulling in worker-only Fetcher types)
- web/next.config.ts: transpilePackages for raw TS from shared package
- web/app/(site)/oauth/consent/page.tsx: replaces local isAllowedRedirectUri with isValidRedirectUri
