+++ web/next.config.ts
@@ -1,6 +1,7 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
+  transpilePackages: ['@flaim/worker-shared'],
   devIndicators: false,
   async redirects() {
     return [
