+++ web/package.json
@@ -12,6 +12,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@flaim/worker-shared": "*",
     "@clerk/nextjs": "^6.36.7",
     "@clerk/themes": "^2.4.51",
     "@radix-ui/react-dialog": "^1.1.14",
